pyproject.toml
[project]
name = "python-project"
version = "0.1.0"
description = "Multi-agent autonomous workflow system for Claude Code"
authors = [
    {name = "bojunC",email = "316231662@qq.com"}
]
requires-python = ">=3.11"
dependencies = [
    "anthropic (>=0.42.0,<1.0.0)",
    "PyGithub (>=2.6.0,<3.0.0)",
    "gitpython (>=3.1.44,<4.0.0)",
    "pydantic (>=2.11.0,<3.0.0)",
    "rich (>=13.9.0,<14.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "tenacity (>=9.0.0,<10.0.0)",
    "httpx (>=0.28.1,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "ruff (>=0.15.7,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)"
]

# Tool configurations

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
show_error_context = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --cov=src/python_project --cov-report=term-missing --cov-report=html"
